The connection between the UI and the API has been finalized.

+++ PeopleViewApp/Services/Interfaces/IUsersApi.cs
@@ -5,9 +5,9 @@ namespace PeopleViewApp.Services.Interfaces
     public interface IUsersApi
     {
         Task<IEnumerable<User>> GetUsers();
-        Task<User> GetUser(int id);
+        Task<User> GetUser(long id);
         Task<User> EditUser(User user);
-        Task DeleteUser(int id);
+        Task DeleteUser(long id);
         Task<User> CreateUser(User user);
 
     }
